requirements.txt
# Database
psycopg2-binary>=2.9.5
asyncpg>=0.27.0
aiosqlite>=0.18.0

# HTTP and async
aiohttp>=3.8.3
requests>=2.28.1
asyncio>=3.4.3

# General utilities
python-dotenv>=0.21.0
tabulate>=0.9.0

# Command-line and formatting
argparse>=1.4.0
tqdm>=4.64.1

# For shodan integration
shodan>=1.28.0

# sqlite3 is part of the Python standard library 